whatsapp-integration-kit/requirements.txt
# WhatsApp Integration Kit - Dependencies

# Core dependencies
httpx>=0.24.0        # Async HTTP client
loguru>=0.7.0        # Advanced logging
python-dotenv>=1.0.0 # Environment variables

# Optional dependencies for examples
fastapi>=0.104.0     # For webhook examples
uvicorn>=0.24.0      # ASGI server
python-multipart     # For file uploads
